requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
SQLAlchemy==2.0.32
typing_extensions==4.12.2
watchdog==4.0.1
Werkzeug==3.0.3
WTForms==3.1.2
